Fix warning message and add redux to HOC
React was throwing warnings about a unknown props to the SideBar.Pusher, after a bit of reading at its docs (and googling), the pusher does not seem to be needed for the overlay behavior we want the sidebar to have. There was no error thrown when I removed the pusher as well.